intermediate update

+++ code/2018/src/main/scala/adventofcode/aoc2018/Day23.scala
@@ -10,7 +10,7 @@ object Day23 extends AdventOfCode(Prod):
       .linesIterator
       .collect:
         case s"pos=<$x,$y,$z>, r=$r" =>
-          Nanobot(Pos3D(x.toLong, y.toLong, z.toLong), r.toLong)
+          Nanobot(Pos3D(x.toInt, y.toInt, z.toInt), r.toInt)
       .toVector
 
   case class Nanobot(pos: Pos3D, signalRadius: Long):
